improve ui index file from 4.7mb to 1.4mb with proper code splitting
