+++ ui/src/resources/server/stats/index.tsx
@@ -1,4 +1,4 @@
-import { ReactNode } from "react";
+import { lazy, ReactNode, Suspense } from "react";
 import { usePermissions, useRead } from "@/lib/hooks";
 import { Types } from "komodo_client";
 import { Section } from "mogh_ui";
@@ -6,9 +6,11 @@ import ServerProcesses from "./processes";
 import ServerContainerStats from "./containers";
 import ServerDisks from "./disks";
 import ServerCurrentStats from "./current";
-import ServerHistoricalStats from "./historical";
 import ServerSystemInfo from "./system-info";
 import { useIsServerAvailable } from "../hooks";
+
+// Loaded lazily to keep recharts out of the entry chunk.
+const ServerHistoricalStats = lazy(() => import("./historical"));
 
 export default function ServerStats({
   id,
@@ -35,7 +37,9 @@ export default function ServerStats({
 
       <ServerCurrentStats id={id} stats={stats} />
 
-      <ServerHistoricalStats id={id} />
+      <Suspense fallback={null}>
+        <ServerHistoricalStats id={id} />
+      </Suspense>
 
       <ServerContainerStats id={id} />
